+++ POS---System-frontend/src/routes/SuperAdminRoutes.jsx
@@ -11,6 +11,7 @@ import StoreListPage from "../pages/SuperAdminDashboard/store/StoreListPage";
 import StoreDetailsPage from "../pages/SuperAdminDashboard/store/StoreDetailsPage";
 import PendingRequestsPage from "../pages/SuperAdminDashboard/store/PendingRequestsPage";
 import SettingsPage from "@/pages/SuperAdminDashboard/settings/SettingsPage";
+import ExportsPage from "@/pages/SuperAdminDashboard/ExportsPage";
 
 const SuperAdminRoutes = () => {
   return (
@@ -22,6 +23,7 @@ const SuperAdminRoutes = () => {
         <Route path="stores/:id" element={<StoreDetailsPage />} />
         <Route path="requests" element={<PendingRequestsPage />} />
         <Route path="subscriptions" element={<SubscriptionPlansPage />} />
+        <Route path="exports" element={<ExportsPage />} />
         <Route path="settings" element={<SettingsPage />} />
       </Route>
     </Routes>